LPS-63096 Rename com.liferay.portal.wab.generator -> com.liferay.portal.osgi.web.wab.generator
