+++ modules/portal-osgi-web/portal-osgi-web-wab-generator/src/main/java/com/liferay/portal/osgi/web/wab/generator/internal/artifact/WarArtifactUrlTransformer.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.wab.generator.internal.artifact;
+package com.liferay.portal.osgi.web.wab.generator.internal.artifact;
 
 import java.io.File;
