+++ modules/portal-osgi-web/portal-osgi-web-wab-generator/src/main/java/com/liferay/portal/osgi/web/wab/generator/internal/handler/WabURLStreamHandlerService.java
@@ -12,25 +12,36 @@
  * details.
  */
 
-package com.liferay.portal.wab.generator.internal.introspection;
+package com.liferay.portal.osgi.web.wab.generator.internal.handler;
 
-import java.io.InputStream;
+import com.liferay.portal.osgi.web.wab.generator.internal.connection.WabURLConnection;
+
+import java.net.URL;
+import java.net.URLConnection;
+
+import org.osgi.framework.BundleContext;
+import org.osgi.service.url.AbstractURLStreamHandlerService;
 
 /**
  * @author Miguel Pastor
  * @author Raymond Augé
  */
-public class ClassLoaderSource implements Source {
+public class WabURLStreamHandlerService
+	extends AbstractURLStreamHandlerService {
 
-	public ClassLoaderSource(ClassLoader classLoader) {
+	public WabURLStreamHandlerService(
+		BundleContext bundleContext, ClassLoader classLoader) {
+
+		_bundleContext = bundleContext;
 		_classLoader = classLoader;
 	}
 
 	@Override
-	public InputStream getResourceAsStream(String name) {
-		return _classLoader.getResourceAsStream(name);
+	public URLConnection openConnection(URL url) {
+		return new WabURLConnection(_bundleContext, _classLoader, url);
 	}
 
+	private final BundleContext _bundleContext;
 	private final ClassLoader _classLoader;
 
 }
