+++ modules/portal-osgi-web/portal-osgi-web-wab-generator/src/main/java/com/liferay/portal/osgi/web/wab/generator/internal/processor/WabProcessor.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.wab.generator.internal.processor;
+package com.liferay.portal.osgi.web.wab.generator.internal.processor;
 
 import aQute.bnd.osgi.Analyzer;
 
@@ -51,9 +51,9 @@ import com.liferay.portal.kernel.xml.SAXReaderUtil;
 import com.liferay.portal.kernel.xml.UnsecureSAXReaderUtil;
 import com.liferay.portal.kernel.xml.XPath;
 import com.liferay.portal.util.PropsValues;
-import com.liferay.portal.wab.generator.internal.introspection.ClassLoaderSource;
-import com.liferay.portal.wab.generator.internal.introspection.Source;
-import com.liferay.portal.wab.generator.internal.util.AntUtil;
+import com.liferay.portal.osgi.web.wab.generator.internal.introspection.ClassLoaderSource;
+import com.liferay.portal.osgi.web.wab.generator.internal.introspection.Source;
+import com.liferay.portal.osgi.web.wab.generator.internal.util.AntUtil;
 import com.liferay.util.xml.XMLUtil;
 
 import java.io.File;
@@ -919,7 +919,7 @@ public class WabProcessor {
 					getVersionedServicePackageName(partialPackageName));
 			}
 
-			_importPackageNames.add("com.liferay.portal.wab.generator");
+			_importPackageNames.add("com.liferay.portal.osgi.web.wab.generator");
 
 			File metaInfDir = new File(uri.resolve("WEB-INF/classes/META-INF"));
 
@@ -930,7 +930,7 @@ public class WabProcessor {
 			FileUtil.write(
 				new File(metaInfDir, "service-bean-post-processor-spring.xml"),
 				clazz.getResourceAsStream(
-					"/com/liferay/portal/wab/generator/internal/dependencies" +
+					"/com/liferay/portal/osgi/web/wab/generator/internal/dependencies" +
 						"/service-bean-post-processor-spring.xml"));
 		}
 		catch (Exception e) {
