+++ modules/portal-osgi-web/portal-osgi-web-wab-generator/src/main/resources/com/liferay/portal/osgi/web/wab/generator/internal/dependencies/service-bean-post-processor-spring.xml
@@ -5,5 +5,5 @@
 	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 	xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.0.xsd"
 >
-	<bean id="com.liferay.portal.wab.generator.ServiceBeanPostProcessor" class="com.liferay.portal.wab.generator.ServiceBeanPostProcessor" />
+	<bean id="com.liferay.portal.osgi.web.wab.generator.ServiceBeanPostProcessor" class="com.liferay.portal.osgi.web.wab.generator.ServiceBeanPostProcessor" />
 </beans>
